Fix an issue with jest & css modules.
For some reason, jest is trying to parase css modules in ALL files. This
means that js files (even if they are UNRELATED to any tests) would cause
jest to crash.

Example line in a js file cauing a crash:
import style from './style.module.scss';

The solution is to tell jest to use a mock CSS Module when a CSS file
is imported.

Since we don't actually care about mocking css files, we just create an
empty file as our mock.

Useful links pertaining to this issue:
- https://stackoverflow.com/questions/39641068/jest-trying-to-parse-css
- https://medium.com/trabe/testing-css-modules-in-react-components-with-jest-enzyme-and-a-custom-modulenamemapper-8ff86c7d18a2